The IIR based limiter runs but...
I don't know if it's working properly: It's output can be very wavy.

We need to check if this is because I now have the signal that is the
convolution with the IRs represent the factor by which the signal is multiplied
and not its divisor. Also the shapes of the curves are a bit different, for
example, with a peak value at sample 1000, -3dB is already reached by sample
300...

Also, we need to properly straighten out the amount of lookahead required.