Limiter now uses lib/datastructures/ringbuffer.c from adaptive_pitch_tracker_extracter
limiter_ir_af can be configured to limit multi-channel signals by
specifying the n_channels parameter.

Starting to add a test (test/stereo/stereo_limiter.c) but it is not
completed.

It has been compiled into libwsola.so in
adaptive_pitch_tracker_extracter and at least doesn't cause problems, so
now it just remains to see that it actually limits.

ALSO TODO: Check to see that the peaks are truely attenuated going from
maximum to minimum.